requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
psycopg2-binary==2.8.5
Flask-Bcrypt==0.7.1
flask-sqlacodegen==1.1.7
flask-babel==1.0.0
Flask-Caching==1.9.0
mailer==0.8.1
PyJWT==1.7.1
python-dateutil==2.8.1
Flask-Limiter==1.3.1
urlquote==1.1.4
Flask-RESTful==0.3.8
webargs==6.1.0
# proxmoxer==1.1.1
paramiko==2.7.1
requests==2.24.0
Flask-HTTPAuth==4.1.0
Flask-Session==0.3.2
redis==3.5.3
cryptography==3.0
APScheduler==3.6.3
sentry-sdk==0.16.3
python-netbox==0.0.15
Flask-Cors==3.0.8
flask-swagger==0.2.14
flask-swagger-ui==3.25.0
Flask-Executor==0.9.3
oslo.serialization==4.0.0
shade==1.33.0 # Apache-2.0
mock==4.0.2
python-openstackclient==5.3.0
python-magnumclient==3.1.0
python-glanceclient==3.2.0
python-neutronclient==7.2.0
python-troveclient==5.0.0
mysqlclient==2.0.1
PyMySQL=0.10.1
